requirements.txt
appdirs==1.4.4
asgiref==3.3.1
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.9.3
bleach==3.1.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.3
comtypes==1.1.10
coverage==5.5
cycler==0.10.0
decorator==4.4.1
defusedxml==0.6.0
deprecation==2.1.0
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.7
django-heroku==0.3.1
entrypoints==0.3
eyed3==0.9.6
filelock==3.0.12
filetype==1.0.7
future==0.18.2
futures==3.1.1
idna==2.10
importlib-metadata==1.2.0
ipykernel==5.1.3
ipython==7.10.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.15.1
Jinja2==2.10.3
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.1
kiwisolver==1.1.0
lxml==4.6.3
MarkupSafe==1.1.1
matplotlib==3.1.2
mistune==0.8.4
more-itertools==8.0.2
nbconvert==5.6.1
nbformat==4.4.0
networkx==2.5
nltk==3.5
notebook==6.0.2
numpy==1.17.4
opencv-contrib-python-headless==4.4.0.42
opencv-python==4.4.0.44
packaging==20.9
pandas==0.25.3
pandocfilters==1.4.2
parso==0.5.1
pickleshare==0.7.5
prometheus-client==0.7.1
prompt-toolkit==2.0.10
psycopg2==2.8.6
Pygments==2.5.2
pyparsing==2.4.5
pypiwin32==223
pyrsistent==0.15.6
PySocks==1.7.1
python-dateutil==2.8.1
python-decouple==3.4
python-whois==0.7.3
python-whois-extended==0.7.12
pythonwhois==2.4.3
pyttsx3==2.90
pytz==2019.3
pywin32==227
pywinpty==0.5.7
pyzmq==18.1.1
qtconsole==4.6.0
regex==2021.3.17
requests==2.25.1
requests-file==1.5.1
scikit-fuzzy==0.4.2
scikit-learn==0.24.1
scipy==1.3.3
seaborn==0.9.0
Send2Trash==1.5.0
six==1.13.0
soupsieve==2.2.1
sqlparse==0.4.1
tabulate==0.8.9
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.1.0
tldextract==3.1.0
toml==0.10.2
tornado==6.0.3
tqdm==4.59.0
traitlets==4.3.3
urllib3==1.26.4
virtualenv==20.4.7
wcwidth==0.1.7
webencodings==0.5.1
whitenoise==5.2.0
whois==0.9.13
widgetsnbextension==3.5.1
zipp==0.6.0
